+++ Communication/Packets/Incoming/Catalog/CheckGnomeNameEvent.cs
@@ -33,7 +33,7 @@ internal class CheckGnomeNameEvent : RoomPacketEvent
 
         var itemId = packet.ReadUInt();
         var item = room.GetRoomItemHandler().GetItem(itemId);
-        if (item == null || item.Definition == null || item.UserId != habbo.Id || item.Definition.InteractionType != InteractionType.GnomeBox)
+        if (item == null || item.Definition == null || item.UserId != habbo.Id || !item.Definition.IsGnomeBox)
             return Task.CompletedTask;
         var petName = packet.ReadString();
         if (string.IsNullOrEmpty(petName))