+++ Communication/Packets/Incoming/Rooms/Furni/Stickys/GetStickyNoteEvent.cs
@@ -10,7 +10,7 @@ internal class GetStickyNoteEvent : RoomPacketEvent
     public override Task Parse(Room room, GameClient session, IIncomingPacket packet)
     {
         var item = room.GetRoomItemHandler().GetItem(packet.ReadUInt());
-        if (item == null || item.Definition.InteractionType != InteractionType.Postit)
+        if (item == null || !item.Definition.IsPostIt)
             return Task.CompletedTask;
         session.Send(new StickyNoteComposer(item.Id.ToString(), item.LegacyDataString));
         return Task.CompletedTask;